secrets: skip generating existing secrets

+++ hack/bin/generate_secrets.sh
@@ -54,6 +54,10 @@ fi
 
 for t in `find "${TARGET_PATH}" -name "*.tpl" -type f`; do
     . "${t}"
+    if [ -f "${TARGET_FILEPATH}" ]; then
+        echo -e "skipping: \"${TARGET_FILEPATH}\" already exists"
+        continue
+    fi
     echo -e "generating sealed-secret at \"${TARGET_FILEPATH}\" with scope \"${TARGET_SCOPE}\" signed by \"${TARGET_CERT}\""
     echo -e "${TEMPLATE}" | kubeseal --name "${SEALED_SECRET_NAME}" --format yaml --cert "${GIT_ROOT}/${TARGET_CERT}" --scope "${TARGET_SCOPE}" > "${GIT_ROOT}/${TARGET_FILEPATH}"
     unset TARGET_CERT; unset TARKET_SCOPE; unset TARGET_FILEPATH; unset TEMPLATE
